Add native desktop shell foundation
Introduce the first native-feeling shell layer for BE Home for Desktop.

This change adds the Tauri-managed app menu, native secondary windows for setup, settings, and about, plus a window-aware React bootstrap that routes the renderer by window label and follows the OS theme.

It also removes the old embedded setup/settings branches from the main workspace, updates the desktop shell tests for the new multi-window model, and wires the official opener plugin/capabilities needed for later support and utility flows.

Validation:
- npm run build
- npm run test
- bounded python ./scripts/dev.py desktop --local-only smoke
