+++ package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@tauri-apps/api": "^2.10.1",
     "@tauri-apps/plugin-dialog": "^2.7.0",
+    "@tauri-apps/plugin-opener": "^2.5.3",
     "react": "^19.1.0",
     "react-dom": "^19.1.0"
   },
