+++ src-tauri/capabilities/default.json
@@ -1,10 +1,13 @@
 {
   "$schema": "../gen/schemas/desktop-schema.json",
   "identifier": "default",
-  "description": "Capability for the main window",
-  "windows": ["main"],
+  "description": "Capability for the desktop shell windows",
+  "windows": ["main", "setup-wizard", "settings", "about"],
   "permissions": [
     "core:default",
-    "dialog:default"
+    "dialog:default",
+    "core:event:default",
+    "core:window:default",
+    "opener:default"
   ]
 }
